+++ core/src/net/oschina/j2cache/J2Cache.java
@@ -15,10 +15,6 @@
  */
 package net.oschina.j2cache;
 
-import net.oschina.j2cache.util.SerializationUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.IOException;
 
 /**
@@ -27,36 +23,14 @@ import java.io.IOException;
  */
 public class J2Cache {
 
-	private final static Logger log = LoggerFactory.getLogger(J2Cache.class);
-
 	private final static String CONFIG_FILE = "/j2cache.properties";
 
-	private final static J2CacheConfig config;
-	private final static CacheChannel channel;
-	private static ClusterPolicy policy; //不同的广播策略
+	private final static J2CacheBuilder builder;
 
 	static {
 		try {
-			config = J2CacheConfig.initFromConfig(CONFIG_FILE);
-			initFromConfig(config);
-			/* 初始化缓存接口 */
-			channel = new CacheChannel(){
-				@Override
-				public void sendClearCmd(String region) {
-					policy.sendClearCmd(region);
-				}
-
-				@Override
-				public void sendEvictCmd(String region, String...keys) {
-					policy.sendEvictCmd(region, keys);
-				}
-
-				@Override
-				public void close() {
-					policy.disconnect();
-					CacheProviderHolder.shutdown();
-				}
-			};
+            J2CacheConfig config = J2CacheConfig.initFromConfig(CONFIG_FILE);
+            builder = J2CacheBuilder.init(config);
 		} catch (IOException e) {
 			throw new CacheException("Failed to load j2cache configuration " + CONFIG_FILE, e);
 		}
@@ -67,27 +41,13 @@ public class J2Cache {
 	 * @return CacheChannel
 	 */
 	public static CacheChannel getChannel(){
-		return channel;
+		return builder.getChannel();
 	}
 
-	/**
-	 * 加载配置
-	 * @return
-	 * @throws IOException
-	 */
-	private static void initFromConfig(J2CacheConfig config) {
-		SerializationUtils.init(config.getSerialization());
-		//初始化两级的缓存管理
-		CacheProviderHolder.init(config, (region, key)->{
-			//当一级缓存中的对象失效时，自动清除二级缓存中的数据
-			CacheProviderHolder.getLevel2Cache(region).evict(key);
-			log.debug(String.format("Level 1 cache object expired, evict level 2 cache object [%s,%s]", region, key));
-			if(policy != null)
-				policy.sendEvictCmd(region, key);
-		});
-
-		policy = ClusterPolicyFactory.init(config.getBroadcast(), config.getBroadcastProperties());
-		log.info("Using cluster policy : " + policy.getClass().getName());
-	}
-
+    /**
+     * 关闭 J2Cache
+     */
+	public static void close() {
+	    builder.close();
+    }
 }